To Do Request Context 적용 / Page Response 적용 / Swagger update

+++ src/main/java/animores/serverapi/to_do/service/impl/ToDoServiceImpl.java
@@ -7,15 +7,15 @@ import animores.serverapi.pet.entity.Pet;
 import animores.serverapi.pet.repository.PetRepository;
 import animores.serverapi.profile.domain.Profile;
 import animores.serverapi.profile.repository.ProfileRepository;
+import animores.serverapi.to_do.dao.GetToDoPageDao;
 import animores.serverapi.to_do.dto.request.ToDoCreateRequest;
 import animores.serverapi.to_do.dto.request.ToDoPatchRequest;
 import animores.serverapi.to_do.dto.response.PetResponse;
 import animores.serverapi.to_do.dto.response.ToDoDetailResponse;
-import animores.serverapi.to_do.dto.response.ToDoResponse;
+import animores.serverapi.to_do.dto.response.ToDoPageResponse;
 import animores.serverapi.to_do.entity.PetToDoRelationship;
 import animores.serverapi.to_do.entity.ToDo;
 import animores.serverapi.to_do.entity.ToDoInstance;
-import animores.serverapi.to_do.entity.vo.ToDoInstanceVo;
 import animores.serverapi.to_do.repository.PetToDoRelationshipRepository;
 import animores.serverapi.to_do.repository.ToDoInstanceRepository;
 import animores.serverapi.to_do.repository.ToDoRepository;
@@ -43,17 +43,7 @@ public class ToDoServiceImpl implements ToDoService {
     @Override
     @Transactional
     public void createToDo(Account account, ToDoCreateRequest request) {
-        Profile createProfile = profileRepository.getReferenceById(2L);
-
-        Set<Long> accountPetIds = petRepository.findAllByAccount_id(account.getId())
-                                                .stream()
-                                                .map(Pet::getId)
-                                                .collect(Collectors.toSet());
-
-        if (!accountPetIds.containsAll(request.petIds())) {
-            throw new CustomException(ExceptionCode.ILLEGAL_PET_IDS);
-        }
-
+        Profile createProfile = profileRepository.getReferenceById(request.profileId());
         ToDo toDo = ToDo.fromRequest(request, account, createProfile);
         toDo = toDoRepository.save(toDo);
         List<PetToDoRelationship> petToDoRelationships = new ArrayList<>();
@@ -69,105 +59,89 @@ public class ToDoServiceImpl implements ToDoService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<ToDoResponse> getTodayToDo(Boolean done, List<Long> pets) {
-
-        Set<Long> petIds = Set.of(1L, 2L);
-        if (pets == null || pets.isEmpty()) {
-            pets = new ArrayList<>(petIds);
-        }
-
-        if (!petIds.containsAll(pets)) {
-            throw new CustomException(ExceptionCode.ILLEGAL_PET_IDS);
-        }
-
-        petRepository.findAllById(pets); //pets 초기화 용
-        List<PetToDoRelationship> relationships = petToDoRelationshipRepository.findAllByPet_IdIn(pets);
+    public ToDoPageResponse getTodayToDo(Boolean done, List<Pet> pets, Integer page, Integer size) {
+        List<Long> petIds = pets.stream().map(Pet::getId).toList();
+        List<PetToDoRelationship> relationships = petToDoRelationshipRepository.findAllByPet_IdIn(petIds);
         //to do id를 key로 하고, 그 to do 에 해당하는 petResonseList를 value로 하는 map
         Map<Long, List<PetResponse>> toDoIdPetResponseMap = relationships.stream()
                 .collect(Collectors.groupingBy(petToDoRelationship -> petToDoRelationship.getToDo().getId(),
                         Collectors.mapping(petToDoRelationship -> new PetResponse(petToDoRelationship.getPet().getId(),
                                 petToDoRelationship.getPet().getName()), Collectors.toList())));
 
-        List<ToDoInstanceVo> toDoInstances;
+        GetToDoPageDao pageDao;
 
         if (done == null) {
-            toDoInstances = toDoInstanceRepository.findAllByTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         } else if (done.equals(Boolean.TRUE)) {
-            toDoInstances = toDoInstanceRepository.findAllByCompleteAndTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByCompleteAndTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         } else {
-            toDoInstances = toDoInstanceRepository.findAllByCompleteFalseAndTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByCompleteFalseAndTodayToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         }
 
-        return toDoInstances.stream()
-                .map(toDoInstanceVo -> ToDoResponse.fromToDoInstanceVo(toDoInstanceVo, toDoIdPetResponseMap.get(toDoInstanceVo.toDo().id())))
-                .toList();
+        return ToDoPageResponse.fromGetToDoPageDaoAndToDoPetMap(pageDao, toDoIdPetResponseMap);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<ToDoResponse> getAllToDo(Boolean done, List<Long> pets) {
-        Set<Long> petIds = Set.of(1L, 2L);
-        if (pets == null || pets.isEmpty()) {
-            pets = petIds.stream().sorted().toList();
-        }
-
-        if (!petIds.containsAll(pets)) {
-            throw new CustomException(ExceptionCode.ILLEGAL_PET_IDS);
-        }
-
-        petRepository.findAllById(pets); //pets 초기화 용
-        List<PetToDoRelationship> relationships = petToDoRelationshipRepository.findAllByPet_IdIn(pets);
+    public ToDoPageResponse getAllToDo(Boolean done, List<Pet> pets, Integer page, Integer size) {
+        List<Long> petIds = pets.stream().map(Pet::getId).toList();
+        List<PetToDoRelationship> relationships = petToDoRelationshipRepository.findAllByPet_IdIn(petIds);
         //to do id를 key로 하고, 그 to do 에 해당하는 petResonseList를 value로 하는 map
         Map<Long, List<PetResponse>> toDoIdPetResponseMap = relationships.stream()
                 .collect(Collectors.groupingBy(petToDoRelationship -> petToDoRelationship.getToDo().getId(),
                         Collectors.mapping(petToDoRelationship -> new PetResponse(petToDoRelationship.getPet().getId(),
                                 petToDoRelationship.getPet().getName()), Collectors.toList())));
 
-        List<ToDoInstanceVo> toDoInstances;
+        GetToDoPageDao pageDao;
         if (done == null) {
-            toDoInstances = toDoInstanceRepository.findAllByToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         } else if (done.equals(Boolean.TRUE)) {
-            toDoInstances = toDoInstanceRepository.findAllByCompleteAndToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByCompleteAndToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         } else {
-            toDoInstances = toDoInstanceRepository.findAllByCompleteFalseAndToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList());
+            pageDao = toDoInstanceRepository.findAllByCompleteFalseAndToDoIdIn(toDoIdPetResponseMap.keySet().stream().toList(), page, size);
         }
 
-        return toDoInstances.stream()
-                .map(toDoInstanceVo -> ToDoResponse.fromToDoInstanceVo(toDoInstanceVo, toDoIdPetResponseMap.get(toDoInstanceVo.toDo().id())))
-                .toList();
+        return ToDoPageResponse.fromGetToDoPageDaoAndToDoPetMap(pageDao, toDoIdPetResponseMap);
 
     }
 
     @Override
     @Transactional(readOnly = true)
-    public ToDoDetailResponse getToDoById(Long id) {
-        ToDo toDo = toDoRepository.findByIdAndAccount_Id(id, 1L).orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
+    public ToDoDetailResponse getToDoById(Long id, Long accountId) {
+        ToDo toDo = toDoRepository.findByIdAndAccount_Id(id, accountId).orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
         return ToDoDetailResponse.fromToDo(toDo);
     }
 
     @Override
     @Transactional
-    public ToDoDetailResponse updateToDoById(Long id, ToDoPatchRequest request) {
-        //TODO: 1L should be replaced with the actual account id
-        ToDo toDo = toDoRepository.findByIdAndAccount_Id(id, 1L).orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
+    public ToDoDetailResponse updateToDoById(Long id, ToDoPatchRequest request, Long accountId) {
+        ToDo toDo = toDoRepository.findByIdAndAccount_Id(id, accountId).orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
 
-        //TODO: 실제 현재 로그인한 사용자의 프로필 아이디로 변경
-        if (toDo.getCreateProfile().getId() != 2L) {
+        if (!toDo.getCreateProfile().getId().equals(request.profileId())) {
             throw new CustomException(ExceptionCode.INAPPROPRIATE_PROFILE_ACCESS);
         }
 
         if (!request.petIds().isEmpty()) {
-            Set<Long> petIds = toDo.getPetToDoRelationships().stream().map(PetToDoRelationship::getPet).map(Pet::getId).collect(Collectors.toSet());
-            if (petIds.size() != request.petIds().size() || !petIds.containsAll(request.petIds())) {
-                petToDoRelationshipRepository.deleteAllByToDo_Id(toDo.getId());
-                List<PetToDoRelationship> petToDoRelationships = new ArrayList<>();
-                for (Long petId : request.petIds()) {
-                    Pet pet = petRepository.getReferenceById(petId);
-                    petToDoRelationships.add(new PetToDoRelationship(pet, toDo));
-                }
-                toDo.setPetToDoRelationships(petToDoRelationships);
-                petToDoRelationshipRepository.saveAll(petToDoRelationships);
-            }
+            Set<Long> petIds = toDo.getPetToDoRelationships().stream()
+                    .map(PetToDoRelationship::getPet)
+                    .map(Pet::getId)
+                    .collect(Collectors.toSet());
+
+            List<Long> petIdsToDelete = petIds.stream().
+                    filter(petId -> !request.petIds().contains(petId))
+                    .toList();
+
+            petToDoRelationshipRepository.deleteAllByToDo_IdAndPet_IdIn(toDo.getId(), petIdsToDelete);
+
+            List<Long> petIdsToAdd = request.petIds().stream()
+                    .filter(petId -> !petIds.contains(petId))
+                    .toList();
+
+            petToDoRelationshipRepository.saveAll(petIdsToAdd.stream()
+                    .map(petRepository::getReferenceById)
+                    .map(pet -> new PetToDoRelationship(pet, toDo))
+                    .toList());
+
         }
 
         toDo.update(request);
@@ -176,13 +150,19 @@ public class ToDoServiceImpl implements ToDoService {
 
     @Override
     @Transactional
-    public void deleteToDoById(Long id) {
+    public void deleteToDoById(Long id, Long profileId) {
+        ToDo todo = toDoRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
+
+        if (!todo.getCreateProfile().getId().equals(profileId)) {
+            throw new CustomException(ExceptionCode.INAPPROPRIATE_PROFILE_ACCESS);
+        }
         toDoRepository.deleteById(id);
     }
 
     @Override
     @Transactional
-    public void checkToDo(Long toDoId) {
+    public void checkToDo(Long toDoId, Long accountId) {
         ToDoInstance toDoInstance = toDoInstanceRepository.findByToDo_IdAndCompleteProfileIsNull(toDoId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_TO_DO));
